+++ packages/frontend/src/components/navigation/ui/NavLinks/index.tsx
@@ -2,39 +2,37 @@ import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { NavLink } from 'react-router-dom';
 
-import { Mixpanel } from '../../../mixpanel/index';
-import ExploreIcon from '../../svg/ExploreIcon';
-import StakingIcon from '../../svg/StakingIcon';
-import UserIcon from '../../svg/UserIcon';
-import WalletIcon from '../../svg/WalletIcon';
-import { StyledLinks, StyledNavItem } from './index';
-
-const track = (msg: string) => Mixpanel.track(msg);
+import { Mixpanel } from '../../../../mixpanel';
+import ExploreIcon from '../../../svg/ExploreIcon';
+import StakingIcon from '../../../svg/StakingIcon';
+import UserIcon from '../../../svg/UserIcon';
+import WalletIcon from '../../../svg/WalletIcon';
+import { StyledLinks, StyledNavItem } from './ui';
 
 const routes = [
     {
-        nameId: 'link.wallet',
+        nameId: 'wallet',
         route: '/',
         trackMsg: 'Click Wallet button on nav',
         icon: <WalletIcon color="var(--navigation-icon-color)" />,
         exact: true,
     },
     {
-        nameId: 'link.staking',
+        nameId: 'staking',
         route: '/staking',
         trackMsg: 'Click Staking button on nav',
         icon: <StakingIcon color="var(--navigation-icon-color)" />,
         testId: 'staking_navlink',
     },
     {
-        nameId: 'link.explore',
+        nameId: 'explore',
         route: '/explore',
         trackMsg: 'Click Explore button on nav',
         icon: <ExploreIcon color="var(--navigation-icon-color)" />,
         testId: 'explore_navlink',
     },
     {
-        nameId: 'link.account',
+        nameId: 'account',
         route: '/profile',
         trackMsg: 'Click Account button on nav',
         icon: <UserIcon color="var(--navigation-icon-color)" />,
@@ -46,6 +44,13 @@ const routes = [
 
 const NavLinks: FC = () => {
     const { t } = useTranslation();
+
+    const translationsMap = {
+        wallet: t('link.wallet'),
+        staking: t('link.staking'),
+        explore: t('link.explore'),
+        account: t('link.account'),
+    };
 
     return (
         <StyledLinks>
@@ -60,18 +65,18 @@ const NavLinks: FC = () => {
                         exact = false,
                         brokenDisplay = false,
                     },
-                    index
+                    index: number
                 ) => (
                     <StyledNavItem key={index}>
                         <NavLink
                             to={route}
-                            onClick={() => track(trackMsg)}
+                            onClick={() => Mixpanel.track(trackMsg)}
                             data-test-id={testId}
                             className={`link ${brokenDisplay ? 'brokenDisplay' : ''}`}
                             exact={exact}
                         >
                             {icon}
-                            <span className="name">{t(nameId)}</span>
+                            <span className="name">{translationsMap[nameId]}</span>
                         </NavLink>
                     </StyledNavItem>
                 )
